refactor: useObservable -> useObservableValue with defaultValue
